bot_service/requirements.txt
aiogram>=3.0.0
google-genai>=0.2.0
python-dotenv>=1.0.0
pymongo>=4.0.0
httpx>=0.20.0